rename ibmq_(local_)emulator file
